+++ lib-semantic-cemantix/src/commonMain/kotlin/fr/outadoc/semantique/api/cemantix/model/NearbyItem.kt
@@ -7,5 +7,5 @@ import kotlinx.serialization.Serializable
 data class NearbyItem(
     val word: String,
     val percentile: Int,
-    val score: Double
+    val score: Float
 )
